requirements.txt
django<2
django-extensions

pylint
pylint-django
autopep8

djangorestframework
simplejson

requests

boto3